feat: 【APM】默认时间切换逻辑优化 --Story=135925238
