+++ bkmonitor/webpack/pnpm-workspace.yaml
@@ -41,6 +41,7 @@ publicHoistPattern:
 
 minimumReleaseAgeExclude:
   - '@blueking/monitor-trace-explore@0.0.21 || 0.0.23 || 0.0.24 || 0.0.25 || 0.0.26 || 0.0.27 || 0.0.28'
+  - '@blueking/open-telemetry@0.0.14'
 
 allowBuilds:
   '@parcel/watcher': set this to true or false
